ajout shnapsho company

+++ src/main/java/org/avricot/rating/service/rule/HashMapNullSafe.java
@@ -5,14 +5,14 @@ import java.util.HashMap;
 /**
  * Return never-null value (0 if null).
  */
-public class HashMapNullSafe extends HashMap<Integer, Integer> {
+public class HashMapNullSafe extends HashMap<Integer, Float> {
     private static final long serialVersionUID = 1L;
 
     @Override
-    public Integer get(final Object key) {
-        Integer value = super.get(key);
+    public Float get(final Object key) {
+        Float value = super.get(key);
         if (value == null) {
-            return 0;
+            return 0F;
         }
         return value;
     }
